solved all problems from comments

+++ src/main/java/edu/hw2/tasks/task2/ImmutableSquare.java
@@ -5,7 +5,7 @@ public class ImmutableSquare extends ImmutableRectangle {
         super(0, 0);
     }
 
-    ImmutableSquare(int size) {
+    ImmutableSquare(double size) {
         super(size, size);
     }
 }
